Show total holidays on first sight
