+++ src/App.js
@@ -1,13 +1,23 @@
-import React, { useState } from 'react';
-import { useSelector } from 'react-redux';
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Home from './pages/home/Home';
 import Header from './components/header/Header';
 import Details from './pages/details/Details';
 import Error404 from './pages/Error404';
+import {
+  fetchEnglandData,
+  fetchIrelandData,
+  fetchScotlandData,
+} from './store/actions/index';
 
 const App = () => {
-  const [total, setTotal] = useState([]);
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(fetchEnglandData());
+    dispatch(fetchIrelandData());
+    dispatch(fetchScotlandData());
+  }, []);
 
   const england = useSelector((state) => state.Reducer.england);
   const ireland = useSelector((state) => state.Reducer.ireland);
@@ -22,37 +32,19 @@ const App = () => {
         <Route
           path="/"
           element={
-            <Home
-              total={total}
-              setTotal={setTotal}
-              england={england}
-              ireland={ireland}
-              scotland={scotland}
-            />
+            <Home england={england} ireland={ireland} scotland={scotland} />
           }
         />
         <Route
           path="/details"
           element={
-            <Details
-              total={total}
-              setTotal={setTotal}
-              england={england}
-              ireland={ireland}
-              scotland={scotland}
-            />
+            <Details england={england} ireland={ireland} scotland={scotland} />
           }
         />
         <Route
           path="/details/:country"
           element={
-            <Details
-              total={total}
-              setTotal={setTotal}
-              england={england}
-              ireland={ireland}
-              scotland={scotland}
-            />
+            <Details england={england} ireland={ireland} scotland={scotland} />
           }
         />
         <Route path="*" element={<Error404 />} />
